# apps/server/test/unit/common/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/payment/plans": {
      "get": {
        "operationId": "PaymentController_listPlans",
        "summary": "套餐列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/payment/orders": {
      "post": {
        "operationId": "PaymentController_createOrder",
        "summary": "创建支付订单",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "PaymentController_listOrders",
        "summary": "支付订单列表",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/payment-config": {
      "get": {
        "operationId": "PaymentController_getConfig",
        "summary": "获取支付配置 [管理员]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "PaymentController_updateConfig",
        "summary": "更新支付配置 [管理员]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/plans": {
      "get": {
        "operationId": "PaymentController_adminPlans",
        "summary": "套餐列表 [管理员]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/plans/{slug}": {
      "put": {
        "operationId": "PaymentController_updatePlan",
        "summary": "更新套餐 [管理员]",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/tenants": {
      "get": {
        "operationId": "PaymentController_adminTenants",
        "summary": "租户列表 [超管]",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/tenants/{id}": {
      "put": {
        "operationId": "PaymentController_adminUpdateTenant",
        "summary": "更新租户 [超管]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/tenants/{id}/subscribe": {
      "post": {
        "operationId": "PaymentController_adminSubscribeTenant",
        "summary": "手动订阅 [超管]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/payment-orders": {
      "get": {
        "operationId": "PaymentController_adminPaymentOrders",
        "summary": "全部支付订单 [超管]",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/payment-orders/{id}/confirm": {
      "post": {
        "operationId": "PaymentController_adminConfirmPayment",
        "summary": "手动确认收款 [超管]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/payment/pay/{tradeNo}/complete": {
      "post": {
        "operationId": "PaymentController_completePayment",
        "summary": "完成支付 (扫码后回调)",
        "parameters": [
          {
            "name": "tradeNo",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/payment/callback/{method}": {
      "post": {
        "operationId": "PaymentController_paymentCallback",
        "summary": "支付回调 (微信/支付宝)",
        "parameters": [
          {
            "name": "method",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "支付订阅"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WXGZH",
    "description": "",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}